Layout in tus vias y refactor in routes
